增加警告与检查
在给http_response中增加content-length字段的时候，发现headers_["content-length"] = (size_t)10可以通过编译。
意识到必须增加类型转换检查，避免这一类奇奇怪怪的现象发生。
详情见http_response.cc

+++ contrib/http/http_request.cc
@@ -102,11 +102,11 @@ void HttpRequest::Process(Buffer *buf) {
         return;
       }
 
-      buf->Pop(target_end_idx + 1);
+      buf->Pop(static_cast<int>(target_end_idx) + 1);
       clrf_idx -= target_end_idx + 1;
       std::string version{buf->begin(), static_cast<size_t>(clrf_idx)};
       assert(0 == memcmp(buf->begin() + clrf_idx, "\r\n", 2));
-      buf->Pop(clrf_idx + 2);
+      buf->Pop(static_cast<int>(clrf_idx) + 2);
       state_ = HttpParseState::kExpectHeaderLine;
     } else if (state_ == HttpParseState::kExpectHeaderLine) {
       // field-line     = field-name ":" OWS field-value OWS
@@ -144,7 +144,7 @@ void HttpRequest::Process(Buffer *buf) {
       std::string_view field_content{line.substr(colon_idx + 1, line.size())};
       field_content = strip(field_content);
       headers_.emplace(std::move(field_name), field_content);
-      buf->Pop(clrf_idx + 2);
+      buf->Pop(static_cast<int>(clrf_idx) + 2);
     } else if (state_ == HttpParseState::kExpectBody) {
       // message-body = *OCTET
       // RFC文档：
@@ -153,7 +153,8 @@ void HttpRequest::Process(Buffer *buf) {
       // Request message framing is independent of method semantics.
       auto content_length_iter = headers_.find("content-length");
       if (content_length_iter != headers_.end()) {
-        if (buf->size() >= std::stoi(content_length_iter->second)) {
+        if (static_cast<ssize_t>(buf->size()) >=
+            std::stoi(content_length_iter->second)) {
           state_ = HttpParseState::kDone;
         }
         return;
