+++ src/buffer.cc
@@ -12,17 +12,17 @@ ssize_t Buffer::ReadFd(int fd, int *err) {
   char extra[65535];
   struct iovec vec[2];
   vec[0].iov_base = data_.data() + tail_;
-  vec[0].iov_len = data_.size() - tail_;
+  vec[0].iov_len = data_.size() - static_cast<size_t>(tail_);
   vec[1].iov_base = extra;
   vec[1].iov_len = sizeof(extra);
   auto ret = ::readv(fd, vec, 2);
   if (ret < 0) {
     *err = errno;
-  } else if (ret <= vec[0].iov_len) {
+  } else if (ret <= static_cast<ssize_t>(vec[0].iov_len)) {
     tail_ += ret;
   } else {
     tail_ += vec[0].iov_len;
-    Append(extra, ret - vec[0].iov_len);
+    Append(extra, static_cast<int>(ret - static_cast<ssize_t>(vec[0].iov_len)));
   }
   Trace("readed {} bytes from fd={}, now head={} tail={} err={} data.size={}",
         ret,
@@ -35,7 +35,7 @@ ssize_t Buffer::ReadFd(int fd, int *err) {
 }
 void Buffer::Shrink() {
   MoveTo(prepend_);
-  data_.resize(tail_);
+  data_.resize(static_cast<size_t>(tail_));
   data_.shrink_to_fit();
 }
 void Buffer::Append(const char *data, int size) {
@@ -46,7 +46,7 @@ void Buffer::Append(const char *data, int size) {
 void Buffer::EnsureWritableSpace(int size) {
   if (WriteableBytes() - prepend_ < size) {
     // 整体空间不够，仅仅只扩容，此处不做内存移动
-    data_.resize(size + tail_ + data_.size());
+    data_.resize(static_cast<size_t>(size + tail_) + data_.size());
     data_.resize(data_.capacity());
   } else {
     // 整体空间够，移动
@@ -58,7 +58,9 @@ void Buffer::MoveTo(int index) {
   if (head_ == index) {
     return;
   }
-  ::memmove(data_.data() + index, data_.data() + head_, tail_ - head_);
+  ::memmove(data_.data() + index,
+            data_.data() + head_,
+            static_cast<size_t>(tail_ - head_));
   tail_ = index + tail_ - head_;
   head_ = index;
 }
@@ -69,15 +71,15 @@ void Buffer::Prepend(const char *data, int size) {
 }
 bool Buffer::StartWith(const char *data, int size) {
   if (ReadableBytes() < size) return false;
-  return 0 == memcmp(data, begin(), size);
+  return 0 == memcmp(data, begin(), static_cast<size_t>(size));
 }
 ssize_t Buffer::FirstOf(const char *data, int size) {
-  void *start = memmem(begin(), this->size(), data, size);
+  void *start = memmem(begin(), this->size(), data, static_cast<size_t>(size));
   return start == NULL ? -1 : static_cast<char *>(start) - begin();
 }
 bool Buffer::Contains(char ch) {
   for (int i = head_; i < tail_; i++) {
-    if (data_[i] == ch) return true;
+    if (data_[static_cast<size_t>(i)] == ch) return true;
   }
   return false;
 }
