+++ tests/buffer_test.cc
@@ -6,19 +6,19 @@
 
 #include <string>
 void Test1() {
-  int prepend = 8;
-  Buffer buf(10, prepend);
+  unsigned prepend = 8;
+  Buffer buf(10, static_cast<int>(prepend));
   std::string s{"abcdefg"};
   buf.Append(s);
   auto p = buf.Data();
-  for (int i = 0; i < s.size(); i++) {
+  for (unsigned i = 0; i < s.size(); i++) {
     assert(p[i + prepend] == s[i]);
   }
   auto q = buf.Data();
   buf.Prepend("12345678", 8);
   s = "12345678abcdefg";
   assert(p == q);
-  for (int i = 0; i < s.size(); i++) {
+  for (unsigned i = 0; i < s.size(); i++) {
     assert(p[i] == s[i]);
   }
 }
